Merged PR 16: fix time issue with time being set to local time
So the JS date object converts any timestring passed to it to local time but we don't want that! We want to display the time as passed (ie. without a conversion).

To fix this I created a small util that strips the timezone info from the ISO8601 string so JS thinks it's the local time and does NO conversion then we manipulate it as desired.

![img](https://puu.sh/sUUf1/9546e6d1cf.png)