Merged PR 58: Dashboard timezone and Search unit test fixes
This PR addresses a few issues in the code.

Issue. The Dashboard `getMoreAppointments()` method was previously not stripping the timezone from the origin Date item, which caused the unit tests to fail when run from a different timezone. The `date-fns` library does not take timezone complications into account when adding and subtracting days, and does not check validity of dates, which results in the rollover causing incorrect results.

Solution. Strip the timezone information from the origin Date item `currentStartDate` in the `getMoreAppointments()` method.

Issue. The Search unit tests were using `shallow()` rendering via enzyme which does not support React's `ref` concept and references to the items were undefined.

Solution. Switching `shallow` to `mount` allows the references to work correctly and gets rid of the `TypeError` errors that popped up.

Issue. The Search unit tests were intermittently passing due to the fact that React's `setState` is asynchronous and `setImmediate` inside of a `setTimeout` with `0` duration block does not always execute in the same number of cycles. There was a _cart before the horse_ situation when the CPU is fast enough to process things before the state was changed.

Solution. Remove assertion of state between stubbed method call.

Normally I would have had the timezone fix as a separate commit, however without all the changes at once, the tests fail, and so the git commit also failed.