Merged PR 80: setup google maps api injection into application boot
In order to use the google maps api, we have to inject a `script` tag into the document.
Since we don't know the api key and may need to get it from a web service data endpoint, we are unable to simply add the tag to the index.html file and must resort to a programatic method of getting the script loaded.

I have built up a utility to get the api key - temporarily using my own api key for dev purposes.
I have added a util to load the google maps api and finally kick off our react application once the api is loaded.