Merged PR 83: added Modal component
This PR is for a basic Modal component -- to be merged as soon as possible to unblock other devs who are waiting for it.

Usage is as follows
```
<Modal
  showModal={true}
  buttons={['Save', 'Cancel']}
  defaultButton={0}
  buttonCallbacks={[this.onSave, this.onCancel]}>
  <div>
    This will be displayed in the Modal
  </div>
</Modal>
```

`showModal = true` will display the modal.
`showModal = false` will hide the modal (`display: none`)

`buttons` is a list of strings for each button to show in the modal
`defaultButton` is the index of which button to be rendered as white text on blue (a default button)
`buttonCallbacks` is a list of functions to call for each button's `onClick` handler

**Note: A followup PR will be made for unit tests for this component.**